--- business/build.gradle
@@ -1,6 +1,0 @@
-subprojects {
-    dependencies {
-        implementation(project(":data:data-api"))
-        implementation(project(":domain"))
-    }
-}
