--- business/business-api/src/main/java/am/cs322/UserService.java
@@ -1,9 +1,0 @@
-package am.cs322;
-
-import am.cs322.model.UserDTO;
-
-public interface UserService {
-
-    public UserDTO createUser(String firstName, String lastName);
-
-}
