--- business/business-impl/build.gradle
@@ -1,4 +1,0 @@
-dependencies {
-    implementation(project(":business:business-api"))
-}
-
