--- data/build.gradle
@@ -1,5 +1,0 @@
-subprojects {
-    dependencies {
-        implementation(project(":domain"))
-    }
-}
