--- data/data-api/src/main/java/am/cs322/UserRepository.java
@@ -1,9 +1,0 @@
-package am.cs322;
-
-import am.cs322.model.User;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends CrudRepository<User, Long> {
-}
