--- domain/src/main/java/am/cs322/model/UserDTO.java
@@ -1,4 +1,0 @@
-package am.cs322.model;
-
-public record UserDTO(String fullName) {
-}
