--- presentation/presentation-api/src/main/java/am/cs322/UserController.java
@@ -1,10 +1,0 @@
-package am.cs322;
-
-import am.cs322.model.CreateUserRequest;
-import am.cs322.model.UserDTO;
-
-public interface UserController {
-
-    public UserDTO createUser(CreateUserRequest request);
-
-}
