--- presentation/presentation-api/src/main/java/am/cs322/model/CreateUserRequest.java
@@ -1,9 +1,0 @@
-package am.cs322.model;
-
-import jakarta.annotation.Nonnull;
-
-public record CreateUserRequest(@Nonnull
-                                String firstName,
-                                @Nonnull
-                                String lastName) {
-}
