--- presentation/presentation-impl/build.gradle
@@ -1,5 +1,0 @@
-dependencies {
-    implementation(project(":presentation:presentation-api"))
-    implementation(project(":business:business-api"))
-}
-
