--- presentation/presentation-impl/src/main/java/am/cs322/UserControllerImpl.java
@@ -1,25 +1,0 @@
-package am.cs322;
-
-import am.cs322.model.CreateUserRequest;
-import am.cs322.model.UserDTO;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/v1/users")
-public class UserControllerImpl implements UserController {
-
-    private final UserService userService;
-
-    public UserControllerImpl(UserService userService) {
-        this.userService = userService;
-    }
-
-    @Override
-    @PostMapping
-    public UserDTO createUser(@RequestBody CreateUserRequest request) {
-        return userService.createUser(request.firstName(), request.lastName());
-    }
-}
