--- webapp/src/main/java/am/cs322/Cs322HomeworkApplication.java
@@ -1,12 +1,0 @@
-package am.cs322;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication(scanBasePackages = {"am.cs322"})
-public class Cs322HomeworkApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(Cs322HomeworkApplication.class, args);
-    }
-}
