+++ webapp/src/main/java/am/cs322/user/data/UserRepository.java
@@ -1,0 +1,10 @@
+package am.cs322.user.data;
+
+
+import am.cs322.user.domain.model.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+}
