+++ webapp/src/main/java/am/cs322/user/domain/model/UserDTO.java
@@ -1,0 +1,4 @@
+package am.cs322.user.domain.model;
+
+public record UserDTO(String fullName) {
+}
