+++ webapp/src/main/java/am/cs322/user/presentation/api/model/CreateUserRequest.java
@@ -1,0 +1,9 @@
+package am.cs322.user.presentation.api.model;
+
+import jakarta.annotation.Nonnull;
+
+public record CreateUserRequest(@Nonnull
+                                String firstName,
+                                @Nonnull
+                                String lastName) {
+}
