+++ webapp/src/main/resources/application.properties
@@ -1,8 +1,12 @@
 spring.application.name=cs322-homework
 server.port=8080
 
-spring.datasource.url=jdbc:h2:mem:testdb
-spring.datasource.driverClassName=org.h2.Driver
-spring.h2.console.enabled=true
+
+
 spring.jpa.hibernate.ddl-auto=create-drop
 spring.jpa.show-sql=true
+
+spring.datasource.url=jdbc:postgresql://localhost:5432/postgres
+spring.datasource.username=postgres
+spring.datasource.password=mysecretpassword
+spring.datasource.driver-class-name=org.postgresql.Driver
